requirements.txt
streamlit
python-docx
PyPDF2
pandas
pdfkit
openpyxl
langchain
google-generativeai
langchain-google-genai
python-dotenv
pydantic
requests
